move get_ids to forest/utils.py
